Added mapping table - STRIDE-OT10-CWE-OPC-ASVS,
Added mapping table - STRIDE-OT10-CWE-OPC-ASVS,, which will help a team to map the STRIDE threats to other frameworks to either find more specific threats or to find mitigations. This enables a team to not have the burden of maintaining Threat and Security Mitigation libraries.These are now based on the standard OWASP mitigation libraries.